Added: Search Functionality for Teachers and Courses
This commit introduces search functionality for teachers and courses in the application. New functions and methods have been implemented to allow users to efficiently search and filter among the list of available teachers and courses. The addition of this functionality enhances the user experience by facilitating the quick and precise location of specific information about teachers and courses.

Changes made:
- Added search functionality for teachers.
- Implemented search functionality for courses.
- User interface enhancements to facilitate the use of the search functionality.